Port linear solver to new structure
